% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_morans_i}
\alias{calculate_morans_i}
\title{Calculate global and local Moran's I for spatial autocorrelation}
\usage{
calculate_morans_i(
  spatial_data,
  max_distance = NULL,
  n_permutations = 100,
  calculate_local = TRUE
)
}
\arguments{
\item{spatial_data}{Data frame with x, y, cluster columns}

\item{max_distance}{Maximum distance for neighbor definition}

\item{n_permutations}{Number of permutations for significance testing}

\item{calculate_local}{Whether to calculate Local Moran's I (LISA)}
}
\value{
List with global results (data frame) and local results (list)
}
\description{
Calculate global and local Moran's I for spatial autocorrelation
}
\examples{
\dontrun{
morans <- calculate_morans_i(spatial_data, max_distance = 1000)
}
}
