% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calculate_pvclust_pvalue}
\alias{calculate_pvclust_pvalue}
\title{Calculate pvclust AU p-value for a split (optimized)}
\usage{
calculate_pvclust_pvalue(data, split_indices, nboot = 10, fast_mode = FALSE)
}
\arguments{
\item{data}{Data matrix}

\item{split_indices}{Indices for the split}

\item{nboot}{Number of bootstrap replicates (reduced for speed)}

\item{fast_mode}{If TRUE, use simplified approximation}
}
\value{
AU p-value
}
\description{
Calculate pvclust AU p-value for a split (optimized)
}
\keyword{internal}
