% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calculate_silhouette_score}
\alias{calculate_silhouette_score}
\title{Calculate silhouette score for clustering}
\usage{
calculate_silhouette_score(data, clusters)
}
\arguments{
\item{data}{Data matrix}

\item{clusters}{Cluster assignments}
}
\value{
Mean silhouette score
}
\description{
Calculate silhouette score for clustering
}
\keyword{internal}
