% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_umap}
\alias{compute_umap}
\title{Compute UMAP embedding}
\usage{
compute_umap(logr_matrix, n_neighbors = 15, min_dist = 0.1)
}
\arguments{
\item{logr_matrix}{LogR matrix}

\item{n_neighbors}{Number of neighbors}

\item{min_dist}{Minimum distance}
}
\value{
Matrix with UMAP coordinates
}
\description{
Compute UMAP embedding
}
\examples{
\dontrun{
umap_coords <- compute_umap(logr_matrix, n_neighbors = 15)
}
}
