% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{create_bulk_from_bins}
\alias{create_bulk_from_bins}
\title{Create bulk profile by aggregating all single-cell bin counts}
\usage{
create_bulk_from_bins(ascat_rda_file, output_dir)
}
\arguments{
\item{ascat_rda_file}{Path to ASCAT result Rda file}

\item{output_dir}{Output directory}
}
\value{
Bulk track data (list with lCTS.tumour)
}
\description{
Create bulk profile by aggregating all single-cell bin counts
}
\examples{
\dontrun{
bulk_track <- create_bulk_from_bins("ascat.Rda", "output/")
}
}
