% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_cluster_colors}
\alias{create_cluster_colors}
\title{Create consistent colorblind-friendly cluster color palette using RColorBrewer}
\usage{
create_cluster_colors(clusters, max_clusters = 24)
}
\arguments{
\item{clusters}{Vector of cluster assignments}

\item{max_clusters}{Maximum number of clusters to support}
}
\value{
Named vector of colors for each cluster ID
}
\description{
Create consistent colorblind-friendly cluster color palette using RColorBrewer
}
\examples{
clusters <- c(1, 1, 2, 2, 3, 3)
colors <- create_cluster_colors(clusters)
}
