% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading.R
\name{create_genomic_grid}
\alias{create_genomic_grid}
\title{Create genomic grid at specified resolution}
\usage{
create_genomic_grid(profile, resolution)
}
\arguments{
\item{profile}{First profile for chromosome info}

\item{resolution}{Bin size in bp}
}
\value{
Data frame with chr, position
}
\description{
Create genomic grid at specified resolution
}
\examples{
\dontrun{
profile <- read_tsv("sample.ASCAT.scprofile.txt")
grid <- create_genomic_grid(profile, resolution = 1e6)
}
}
