% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading.R
\name{load_ascat_profiles}
\alias{load_ascat_profiles}
\title{Load ASCAT.sc profile files and create matrices}
\usage{
load_ascat_profiles(
  profile_dir,
  pattern = "*refitted.ASCAT.scprofile.txt",
  resolution = 1e+06,
  logr_missing_value = 0
)
}
\arguments{
\item{profile_dir}{Directory containing profile files}

\item{pattern}{File pattern (default: "*refitted.ASCAT.scprofile.txt")}

\item{resolution}{Genomic resolution in bp}

\item{logr_missing_value}{Value for missing logR data}
}
\value{
List with copy_number_matrix, logr_matrix, grid, sample_names
}
\description{
Load ASCAT.sc profile files and create matrices
}
\examples{
\dontrun{
profiles <- load_ascat_profiles("path/to/profiles", resolution = 1e6)
}
}
