% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-matching.R
\name{load_geojson_tiles}
\alias{load_geojson_tiles}
\title{Load and process GeoJSON tile data with spatial coordinates}
\usage{
load_geojson_tiles(geojson_file)
}
\arguments{
\item{geojson_file}{Path to GeoJSON file}
}
\value{
sf object with tile geometries and centroids
}
\description{
Load and process GeoJSON tile data with spatial coordinates
}
\examples{
\dontrun{
tiles <- load_geojson_tiles("path/to/tiles.geojson")
}
}
