% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{merge_bams_by_cluster}
\alias{merge_bams_by_cluster}
\title{Merge BAM files by cluster}
\usage{
merge_bams_by_cluster(
  bam_dir,
  sample_names,
  clusters,
  output_dir,
  min_samples = 3
)
}
\arguments{
\item{bam_dir}{Directory containing BAM files}

\item{sample_names}{Sample names}

\item{clusters}{Cluster assignments}

\item{output_dir}{Output directory for merged BAMs}

\item{min_samples}{Minimum samples per cluster}
}
\value{
Character vector of merged BAM file paths
}
\description{
Merge BAM files by cluster
}
\examples{
\dontrun{
merged_bams <- merge_bams_by_cluster("bams/", sample_names, clusters, "merged/")
}
}
