% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{merge_small_clusters}
\alias{merge_small_clusters}
\title{Merge small clusters into nearest large clusters}
\usage{
merge_small_clusters(clusters, data, min_size)
}
\arguments{
\item{clusters}{Vector of cluster assignments}

\item{data}{Data matrix}

\item{min_size}{Minimum cluster size threshold}
}
\value{
Vector of merged cluster assignments
}
\description{
Merge small clusters into nearest large clusters
}
\keyword{internal}
