% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_chromosomal_instability}
\alias{plot_chromosomal_instability}
\title{Plot chromosomal instability index}
\usage{
plot_chromosomal_instability(copy_number_matrix, grid, output_file)
}
\arguments{
\item{copy_number_matrix}{Copy number matrix}

\item{grid}{Genomic grid}

\item{output_file}{Output PDF path}
}
\description{
Plot chromosomal instability index
}
\examples{
\dontrun{
plot_chromosomal_instability(cn_matrix, grid, "instability.pdf")
}
}
