% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_cluster_correlation}
\alias{plot_cluster_correlation}
\title{Plot cluster correlation heatmap}
\usage{
plot_cluster_correlation(logr_matrix, clusters, output_file)
}
\arguments{
\item{logr_matrix}{LogR matrix}

\item{clusters}{Cluster assignments}

\item{output_file}{Output PDF path}
}
\description{
Plot cluster correlation heatmap
}
\examples{
\dontrun{
plot_cluster_correlation(logr_matrix, clusters, "correlation.pdf")
}
}
