% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_lisa}
\alias{plot_lisa}
\title{Plot Local Indicators of Spatial Association (LISA) results}
\usage{
plot_lisa(lisa_results, tiles, output_file)
}
\arguments{
\item{lisa_results}{List of data frames with Local Moran's I results}

\item{tiles}{GeoJSON tiles for spatial context}

\item{output_file}{Output PDF path}
}
\value{
Summary table of LISA results
}
\description{
Plot Local Indicators of Spatial Association (LISA) results
}
\examples{
\dontrun{
summary <- plot_lisa(lisa_results, tiles, "lisa.pdf")
}
}
