% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_logr_variance}
\alias{plot_logr_variance}
\title{Plot LogR variance per sample}
\usage{
plot_logr_variance(logr_matrix, output_file)
}
\arguments{
\item{logr_matrix}{LogR matrix}

\item{output_file}{Output PDF path}
}
\description{
Plot LogR variance per sample
}
\examples{
\dontrun{
plot_logr_variance(logr_matrix, "variance.pdf")
}
}
