% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ploidy_distribution}
\alias{plot_ploidy_distribution}
\title{Plot total ploidy distribution}
\usage{
plot_ploidy_distribution(copy_number_matrix, output_file)
}
\arguments{
\item{copy_number_matrix}{Copy number matrix}

\item{output_file}{Output PDF path}
}
\description{
Plot total ploidy distribution
}
\examples{
\dontrun{
plot_ploidy_distribution(cn_matrix, "ploidy.pdf")
}
}
