% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_spatial_coherence}
\alias{plot_spatial_coherence}
\title{Plot global spatial coherence using Moran's I statistics}
\usage{
plot_spatial_coherence(morans_results, output_file)
}
\arguments{
\item{morans_results}{Results from calculate_morans_i}

\item{output_file}{Output PDF path}
}
\description{
Plot global spatial coherence using Moran's I statistics
}
\examples{
\dontrun{
plot_spatial_coherence(morans_results, "coherence.pdf")
}
}
