% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot UMAP with cluster colors}
\usage{
plot_umap(umap_coords, clusters, output_file)
}
\arguments{
\item{umap_coords}{UMAP coordinates matrix}

\item{clusters}{Cluster assignments}

\item{output_file}{Output PDF path}
}
\description{
Plot UMAP with cluster colors
}
\examples{
\dontrun{
plot_umap(umap_coords, clusters, "umap.pdf")
}
}
