% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{refine_clusters_by_correlation}
\alias{refine_clusters_by_correlation}
\title{Merge similar clusters based on correlation}
\usage{
refine_clusters_by_correlation(logr_matrix, clusters, cor_threshold = 0.9)
}
\arguments{
\item{logr_matrix}{LogR matrix}

\item{clusters}{Initial cluster assignments}

\item{cor_threshold}{Correlation threshold for merging (default: 0.9)}
}
\value{
Refined cluster assignments
}
\description{
Merge similar clusters based on correlation
}
\examples{
\dontrun{
refined <- refine_clusters_by_correlation(logr_matrix, clusters, cor_threshold = 0.9)
}
}
