% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat-core.R
\name{run_ascat_on_pseudobulk_tracks}
\alias{run_ascat_on_pseudobulk_tracks}
\title{Run full ASCAT.sc on pseudobulk tracks (mirrors run_sc_sequencing workflow)}
\usage{
run_ascat_on_pseudobulk_tracks(pseudobulk_tracks, ascat_rda_file, output_dir)
}
\arguments{
\item{pseudobulk_tracks}{List of pseudobulk track data}

\item{ascat_rda_file}{Path to original ASCAT result}

\item{output_dir}{Output directory}
}
\value{
ASCAT result object with profiles
}
\description{
Run full ASCAT.sc on pseudobulk tracks (mirrors run_sc_sequencing workflow)
}
\examples{
\dontrun{
pb_res <- run_ascat_on_pseudobulk_tracks(pb_tracks, "ascat.Rda", "output")
}
}
